T17 - CrudEntity, Service y Repository de la entidad Beneficiarios

+++ src/main/java/org/albor/fentfood/dominio/dto/BeneficiarioDto.java
@@ -3,7 +3,7 @@ package org.albor.fentfood.dominio.dto;
 import org.albor.fentfood.persistence.entity.BeneficiarioEntity.TipoBeneficiario;
 
 public record BeneficiarioDto(
-        Integer idBeneficiario,
+        Long idBeneficiario,
         String name,
         String lastName,
         String phone,
