+++ src/main/java/org/albor/fentfood/persistence/crud/CrudBeneficiarioEntity.java
@@ -3,6 +3,6 @@ package org.albor.fentfood.persistence.crud;
 import org.albor.fentfood.persistence.entity.BeneficiarioEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CrudBeneficiarioEntity extends CrudRepository<BeneficiarioEntity, Integer> {
-    BeneficiarioEntity findFirstByIdBeneficiario(Integer idBeneficiario);
+public interface CrudBeneficiarioEntity extends CrudRepository<BeneficiarioEntity, Long> {
+    BeneficiarioEntity findFirstByIdBeneficiario(Long idBeneficiario);
 }
